+++ src/Sulu/Bundle/TestBundle/Kernel/SuluTestKernel.php
@@ -32,6 +32,7 @@ class SuluTestKernel extends SuluKernel
             new \Symfony\Bundle\MonologBundle\MonologBundle(),
             new \Symfony\Bundle\SecurityBundle\SecurityBundle(),
             new \Symfony\Bundle\TwigBundle\TwigBundle(),
+            new \Dubture\FFmpegBundle\DubtureFFmpegBundle(),
 
             // Sulu
             new \Sulu\Bundle\CoreBundle\SuluCoreBundle(),
